# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todo Application REST-API scheme",
        "title": "Golang Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "Перегляд списку задач з опціональною пагінацією та/або фільрацією по ID автора задачі",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Cписок задач",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Фільтрація задач по ID автора",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Розмір сторінки з задачами",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Зміщення сторінки з задачами",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_tasks_transport_http.TaskDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Створити нову задачу в системі",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Створити задачу",
                "parameters": [
                    {
                        "description": "CreateTask тіло запиту",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успішно створена задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Отримання конкретної задачі по її ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Отримання задачі",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID отримуваної задачі",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Видалити існуючу в системі задачу по її ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Видалення задачі",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID видаляємої задачі",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успішне видалення задачі"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Оновлює інформацію вже в існуючій в системі задачі\n### Логіка оновлення полів (Three-state-logic)\n1. **Поле не передано**: `description` ігноруєтся, значення в БД не змінюєтся\n2. **Явно передано значення**: `description`: \"Вийти погуляти в 6:30 з собакою\" - змінюєьтся поле в БД\nОбмеження: `title` і `completed` не можуть бути виставденні null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Оновити задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID оновлюваної задачі",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchTask тіло запиту",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PatchTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно змінена задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PatchTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Mirwinli_golang-todoapp_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable"
                }
            }
        },
        "github_com_Mirwinli_golang-todoapp_internal_core_transport_http_types.Nullable-bool": {
            "type": "object",
            "properties": {
                "set": {
                    "type": "boolean"
                },
                "value": {
                    "type": "boolean"
                }
            }
        },
        "github_com_Mirwinli_golang-todoapp_internal_core_transport_http_types.Nullable-string": {
            "type": "object",
            "properties": {
                "set": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "internal_features_tasks_transport_http.CreateTaskRequest": {
            "type": "object",
            "required": [
                "author_user_id",
                "title"
            ],
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1,
                    "example": "Сробити домашку до 16 квітня"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Hello Wo"
                }
            }
        },
        "internal_features_tasks_transport_http.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 1
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:10:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "This is a test description"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Task title"
                },
                "version": {
                    "type": "integer",
                    "example": 1007
                }
            }
        },
        "internal_features_tasks_transport_http.PatchTaskRequest": {
            "type": "object",
            "properties": {
                "completed": {
                    "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_types.Nullable-bool"
                },
                "description": {
                    "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_types.Nullable-string"
                },
                "title": {
                    "$ref": "#/definitions/github_com_Mirwinli_golang-todoapp_internal_core_transport_http_types.Nullable-string"
                }
            }
        },
        "internal_features_tasks_transport_http.PatchTaskResponse": {
            "type": "object",
            "properties": {
                "authorUserID": {
                    "type": "integer"
                },
                "completed": {
                    "type": "boolean"
                },
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "internal_features_tasks_transport_http.TaskDTOResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 1
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:10:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "This is a test description"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Task title"
                },
                "version": {
                    "type": "integer",
                    "example": 1007
                }
            }
        },
...
    }
}